+++ Gainy/Gainy/Scenes/Portfolio/Plaid/Models/HoldingsModelMapper.swift
@@ -119,8 +119,8 @@ struct HoldingsModelMapper {
         let (main, median) = normalizeCharts(chartsCache[range] ?? [ChartNormalized](), spChart)
         
         var chartDataSP: ChartData!
-        if let firstMedian: Float = median.first?.val, let firstMain: Float = main.first?.val {
-            var pcts: [Float] = []
+        if let firstMedian: Double = median.first?.val, let firstMain: Double = main.first?.val {
+            var pcts: [Double] = []
             for val in median.compactMap({$0.val}) {
                 let cur = val / firstMedian
                 pcts.append(firstMain * cur)
